refactor(scripture): book name lookup
Using a cache of user values is dangerous as it could potentially grow infinitely.
The old code also required O(n^2) for non-cached names.

Now we store a map of all possible case-insensitive names to their index.
This has a safe, constant map size and only needs to be done once.
This results in lookup performance of O(1).