fix(scripture): mangling of Book/Chapter/Verse class names in build
Just to provide better DX.
We could enable esbuild's "keep names" option,
but that applies to everything, and we don't really need that.
They have it disabled by default, and I'm assuming there is a
bundle size to consider.
Since this does go to browsers, since is a consideration.
So I'm manually applying to only these 3 cases.