feat(email)!: switch to nodemailer for message compilation and transport
I believe this compilation implementation is more robust
than `emailjs`, and the headers are better typed.
Since it already supports SES v2, there is no reason to not use it for transport as well.

Though the transporter is now injectable.
I adapted its logger (bunyan) to one for NestJS too.