feat(email)!: unify react render for text/html split
Now we only render once for React, meaning that components should
only be rendered once. This is more important now that components
can be async and do data loading work.

The context/hooks for "in text only" have been removed.
Components now need to declarativity state that with data attributes.
This is the biggest breaking change here, but I think it is worth it.
Declaring output for both formats at once is very similar to CSS styling,
giving up control of when each of those outputs is used.
Also, I don't think this is that big of a deal. In practice, we didn't
have any JS logic based on this output format.

We then post-process those data attributes from the HTML given to us
by React to split the output for text & HTML.

Helpers <InText> & <InHtml> can be used to ease this.
Or `data-render-only="text/html"` can be declared anywhere.