feat(common): keys (#69)
I initially tried to avoid this function, because I was hopeful that `entries` would be a singular function that would handle both cases.
I've concluded though that that strategy can be too verbose and unintuitive to read.

So here's the small helper that thinly wraps `entries` and maps to just the key.